Crate op_alloy_consensus

op-alloy-consensus

OP Stack consensus interface.

This crate contains constants, types, and functions for implementing Optimism EL consensus and communication. This includes an extended OpTxEnvelope type with deposit transactions, and receipts containing OP Stack specific fields (deposit_nonce + deposit_receipt_version).

In general a type belongs in this crate if it exists in the alloy-consensus crate, but was modified from the base Ethereum protocol in the OP Stack. For consensus types that are not modified by the OP Stack, the alloy-consensus types should be used instead.

Provenance

Much of this code was ported from reth-primitives as part of ongoing alloy migrations.

Modules

receipt 🔒

transaction 🔒

Structs

OpDepositReceipt

Receipt containing result of transaction execution.

OpDepositReceiptWithBloom

OpDepositReceipt with calculated bloom filter, modified for the OP Stack.

TxDeposit

Deposit transactions, also known as deposits are initiated on L1, and executed on L2.

Enums

OpReceiptEnvelope

Receipt envelope, as defined in EIP-2718, modified for OP Stack chains.

OpTxEnvelope

The Ethereum EIP-2718 Transaction Envelope, modified for OP Stack chains.

OpTxType

Optimism TransactionType flags as specified in EIPs 2718, 1559, and 2930, as well as the deposit transaction spec

OpTypedTransaction

The TypedTransaction enum represents all Ethereum transaction request types, modified for the OP Stack.

Constants

DEPOSIT_TX_TYPE_ID

Identifier for an Optimism deposit transaction

Traits

OpTxReceipt

Receipt is the result of a transaction execution.