op_alloy_consensus::hardforks
Module fjord

Module containing a Transaction builder for the Fjord network upgrade transactions.

Constants

ENABLE_FJORD_FUNC_SIGNATURE

Input data for setting the Fjord Gas Price Oracle.

FJORD_GAS_PRICE_ORACLE_ADDRESS

Fjord Gas Price Oracle address.

GAS_PRICE_ORACLE_ADDRESS

The Gas Price Oracle Address This is computed by using go-ethereum’s crypto.CreateAddress function, with the Gas Price Oracle Deployer Address and nonce 0.

GAS_PRICE_ORACLE_FJORD_DEPLOYER_ADDRESS

Fjord Gas Price Oracle Deployer Address.

L1_INFO_DEPOSITER_ADDRESS

The L1 Info Depositer Address.

SET_FJORD_METHOD_SIGNATURE

The Set Fjord Four Byte Method Signature.

Statics

DEPLOY_FJORD_GAS_PRICE_ORACLE_SOURCE 🔒

Fjord Gas Price Oracle source.

ENABLE_FJORD_SOURCE 🔒

UpgradeDepositSource for setting the Fjord Gas Price Oracle.

UPDATE_FJORD_GAS_PRICE_ORACLE_SOURCE 🔒

UpgradeDepositSource for the source code update to the Fjord Gas Price Oracle.