op_alloy_consensus

Struct OpDepositReceipt

pub struct OpDepositReceipt<T = Log> {
    pub inner: Receipt<T>,
    pub deposit_nonce: Option<u64>,
    pub deposit_receipt_version: Option<u64>,
}

Receipt containing result of transaction execution.

Fields
inner: Receipt<T>

The inner receipt type.

deposit_nonce: Option<u64>

Deposit nonce for Optimism deposit transactions

deposit_receipt_version: Option<u64>

Deposit receipt version for Optimism deposit transactions

The deposit receipt version was introduced in Canyon to indicate an update to how receipt hashes should be computed when set. The state transition process ensures this is only set for post-Canyon deposit transactions.

Implementations

impl OpDepositReceipt

pub fn bloom_slow(&self) -> Bloom

Calculates Log’s bloom filter. this is slow operation and OpDepositReceiptWithBloom can be used to cache this value.

pub fn with_bloom(self) -> OpDepositReceiptWithBloom

Calculates the bloom filter for the receipt and returns the OpDepositReceiptWithBloom container type.

impl<T> OpDepositReceipt<T>

pub fn map_inner<U, F>(self, f: F) -> OpDepositReceipt<U>
where F: FnOnce(Receipt<T>) -> Receipt<U>,

Maps the inner receipt value of this receipt.

This is mainly useful for mapping the receipt log type to the rpc variant.

pub const fn with_bloom_unchecked(self, bloom: Bloom) -> ReceiptWithBloom<Self>

Attaches the given bloom to the receipt returning ReceiptWithBloom.

pub fn into_inner(self) -> Receipt<T>

Consumes the type and returns the inner Receipt.

pub const fn deposit_info(&self) -> OpDepositInfo

Returns the deposit info for this receipt.

pub fn map_logs<U>(self, f: impl FnMut(T) -> U) -> OpDepositReceipt<U>

Converts the receipt’s log type by applying a function to each log.

Returns the receipt with the new log type

impl<T: Encodable> OpDepositReceipt<T>

pub fn rlp_encoded_fields_length_with_bloom(&self, bloom: &Bloom) -> usize

Returns length of RLP-encoded receipt fields with the given Bloom without an RLP header.

pub fn rlp_encode_fields_with_bloom(&self, bloom: &Bloom, out: &mut dyn BufMut)

RLP-encodes receipt fields with the given Bloom without an RLP header.

pub fn rlp_header_with_bloom(&self, bloom: &Bloom) -> Header

Returns RLP header for this receipt encoding with the given Bloom.

impl<T: Decodable> OpDepositReceipt<T>

pub fn rlp_decode_fields_with_bloom( buf: &mut &[u8], ) -> Result<ReceiptWithBloom<Self>>

RLP-decodes receipt’s field with a Bloom.

Does not expect an RLP header.