op_alloy_consensus::transaction

Trait DepositTransaction

pub trait DepositTransaction: Transaction {
    // Required methods
    fn source_hash(&self) -> Option<B256>;
    fn mint(&self) -> u128;
    fn is_system_transaction(&self) -> bool;
}

A trait representing a deposit transaction with specific attributes.

Required Methods

fn source_hash(&self) -> Option<B256>

Returns the hash that uniquely identifies the source of the deposit.

Returns

An Option<B256> containing the source hash if available.

fn mint(&self) -> u128

Returns the optional mint value of the deposit transaction.

Returns

An u128 representing the ETH value to mint on L2, if any.

fn is_system_transaction(&self) -> bool

Indicates whether the transaction is exempt from the L2 gas limit.

Returns

A bool indicating if the transaction is a system transaction.