op_alloy_consensus

Type Alias OpDepositReceiptWithBloom

pub type OpDepositReceiptWithBloom<T = Log> = ReceiptWithBloom<OpDepositReceipt<T>>;

OpDepositReceipt with calculated bloom filter, modified for the OP Stack.

This convenience type allows us to lazily calculate the bloom filter for a receipt, similar to Sealed.

Aliased Type

struct OpDepositReceiptWithBloom<T = Log> {
    pub receipt: OpDepositReceipt<T>,
    pub logs_bloom: Bloom,
}

Fields
receipt: OpDepositReceipt<T>

The receipt.

logs_bloom: Bloom

The bloom filter.